fix: fixed shutdown of binary
